Update project4.erl
Made a rough outline for how a user will interact with the server. Will need to implement logging in and maintaining username/ password so that looking up by mentions works, and the user will remain logged in... Will also need to think about logging out AND writing data to files on quit and reading them on start?